+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Version 0.2.0 (unreleased)
 * Deeper byline implementation
+* Add `primary?` and `standard?` methods to Image to check type
 
 ### Version 0.1.1 (2012-12-16)
 * First stable release.